shtimi i te dhenave te userit.

+++ simora/src/main/java/unipr/fshmn/simora/db/RoomRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called RoomRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface RoomRepository extends CrudRepository<Room, Integer> {
+public interface RoomRepository extends CrudRepository<Room, Long> {
 
 }
